Fix aegis check for solar proximity
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5446 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/OOSunEntity.m
@@ -532,7 +532,7 @@ MA 02110-1301, USA.
 }
 
 
-- (GLfloat) radius
+- (double) radius
 {
 	return collision_radius;
 }
